// IamFuture-Dex-Frontend/src/config/constants/types.ts
export type PageBreadcrumb = {
  title: string
  subtitle?: string
  description?: string
}